Actually make .jenkins_script.sh aware of $HPXIBDEV.
Former-commit-id: ea1d6a2cc1edfaa478b66093b7862031e1733d71

+++ .jenkins_script.sh
@@ -22,15 +22,31 @@ function add_photon() {
 }
 
 set -xe
-export PSM_MEMORY=large
+
+case "$HPXIBDEV" in
+    qib0)
+        export PSM_MEMORY=large
+	TESTCMD="mpirun -np 2 --mca btl_openib_if_include $HPXIBDEV"
+	;;
+    mlx4_0)
+	TESTCMD="mpirun -np 2 --mca mtl ^psm --mca btl_openib_if_include mlx4_0"
+	;;
+    none)
+        TESTCMD=""
+        ;;
+    *)
+	TESTCMD=""
+	;;
+esac
+
 case "$HPXMODE" in
     photon)
-	CFGFLAGS=" --with-mpi=ompi --enable-photon "
+	CFGFLAGS=" --with-mpi=ompi --enable-photon $TESTCMD"
 	add_mpi
         add_photon
 	;;
     mpi)
-	CFGFLAGS=" --with-mpi=ompi "
+	CFGFLAGS=" --with-mpi=ompi $TESTCMD"
 	add_mpi	
 	;;
     *)
@@ -42,11 +58,22 @@ echo "Building HPX in $DIR"
 cd $DIR
 
 module load gcc/4.9.2
-rm -rf ./build/
+
+echo "Bootstrapping HPX."
 ./bootstrap
+
+if [ -d "./build" ]; then
+        rm -rf ./build/
+fi
 mkdir build
 cd build
-../configure --prefix=$DIR/build/HPX5/ $CFGFLAGS --enable-testsuite $HPXDEBUG
+
+if [ -d "./install" ]; then
+        rm -rf ./install/
+fi
+mkdir install
+
+../configure --prefix=$DIR/build/install/ $CFGFLAGS --enable-testsuite $HPXDEBUG
 make
 make install
